refactor: Darkモードしか使わないのでテーマの定義を修正
