+++ pubspec.yaml
@@ -24,7 +24,6 @@ dependencies:
   # localDB
   hive: ^2.2.3
   hive_flutter: ^1.1.0
-  shared_preferences: ^2.2.0
 
   # Firebase
   google_sign_in: ^6.1.4
